try a new ko forbidding method

+++ src/board.h
@@ -359,7 +359,7 @@ public:
 	// AQ-PS
 	// score penalty if a ko is played. used to discorrage kos.
 	//double ko_penalty;
-	double ko_penalty_my;
+	int ko_penalty_my;
 	int searchdepth;
 
 	Board();
